% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_r2.r
\name{AdaBoost.R2.predict}
\alias{AdaBoost.R2.predict}
\title{AdaBoost.R2 predict method}
\usage{
AdaBoost.R2.predict(models, betas, newdata)
}
\arguments{
\item{models}{A set of decision trees corresponding to the number of boosting iterations}

\item{newdata}{A data frame containing the test data}
}
\description{
AdaBoost.R2 predict method
}
\keyword{internal}
