% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AdaBoost.R2}
\alias{AdaBoost.R2}
\title{AdaBoost.R2}
\usage{
AdaBoost.R2(form, train, test, t_final = 100, power = 2, ...)
}
\arguments{
\item{form}{The model formula.}

\item{train}{A data.frame with the training data.}

\item{test}{A data.frame with the test data.}

\item{t_final}{The number of maximum boosting iterations. Default is 100.}

\item{power}{Type of loss function, e.g. linear (1), squared (2). Default is 2.}

\item{...}{Dots are passed to rpart}
}
\value{
Returns a vector with the predictions made by AdaBoost.R2.
}
\description{
AdaBoost.R2
}
\examples{
data(Boston,package="MASS")

idx <- sample(1:nrow(Boston),nrow(Boston)*0.75)
form <- medv ~ .

train <- Boston[idx,]
test <- Boston[-idx,]

preds <- AdaBoost.R2(form,train,test)

}
\references{

}
