% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_r2.r
\name{AdaBoost.R2.train}
\alias{AdaBoost.R2.train}
\title{Create an AdaBoost.R2 model}
\usage{
AdaBoost.R2.train(form, data, niter = 100, power = 2)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{power}{Exponent used to calculate the weight of each decision tree. Common values are 1 (linear), 2 (quadratic, default), 3 (cubic).}
}
\description{
Create an AdaBoost.R2 model
}
\keyword{internal}
