% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rq.r
\name{AdaBoost.RQ.predict}
\alias{AdaBoost.RQ.predict}
\title{AdaBoost.RQ predict method}
\usage{
AdaBoost.RQ.predict(models, newdata, betas)
}
\arguments{
\item{models}{A set of decision trees corresponding to the number of boosting iterations}

\item{newdata}{A data frame containing the test data}

\item{betas}{Model weights}
}
\description{
AdaBoost.RQ predict method
}
\keyword{internal}
