% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rq.r
\name{AdaBoost.RQ.train}
\alias{AdaBoost.RQ.train}
\title{AdaBoost.RQ}
\usage{
AdaBoost.RQ.train(form, data, niter = 100, power = 2)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{niter}{boosting iters}

\item{power}{power??}
}
\description{
AdaBoost.RQ
}
