% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rt.r
\name{AdaBoost.RT.predict}
\alias{AdaBoost.RT.predict}
\title{AdaBoost.RT predict method}
\usage{
AdaBoost.RT.predict(models, newdata, betas)
}
\arguments{
\item{models}{A set of decision trees corresponding to the number of boosting iterations}

\item{newdata}{A data frame containing the test data}

\item{betas}{Model weights}
}
\description{
AdaBoost.RT predict method
}
\keyword{internal}
