% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rt.r
\name{AdaBoost.RT.train}
\alias{AdaBoost.RT.train}
\title{AdaBoost.RT}
\usage{
AdaBoost.RT.train(form, data, niter = 100, thr = 0.1, power = 2)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{thr}{Absolute Relative Error threshold used to define prediction cases as hard to predict}

\item{power}{Exponent used to calculate the weight of each decision tree. Common values are 1 (linear), 2 (quadratic, default), 3 (cubic).}
}
\description{
AdaBoost.RT
}
\keyword{internal}
