% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rtplus.r
\name{AdaBoost.RTPlus.predict}
\alias{AdaBoost.RTPlus.predict}
\title{predict method for AdaBoost.RTPlus}
\usage{
AdaBoost.RTPlus.predict(models, newdata, delta)
}
\arguments{
\item{models}{A set of decision trees corresponding to the number of boosting iterations}

\item{newdata}{A data frame containing the test data}

\item{delta}{Model weights}
}
\description{
predict method for AdaBoost.RTPlus
}
\keyword{internal}
