% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AdaBoost.RTPlus}
\alias{AdaBoost.RTPlus}
\title{AdaBoost.RTPlus}
\usage{
AdaBoost.RTPlus(
  form,
  train,
  test,
  t_final = 100,
  thr = 0.01,
  power = 2,
  sigma = 0.5,
  ...
)
}
\arguments{
\item{form}{The model formula.}

\item{train}{A data.frame with the training data.}

\item{test}{A data.frame with the test data.}

\item{t_final}{The number of maximum boosting iterations. Default is 100.}

\item{thr}{The error threshold. Default is 0.01.}

\item{power}{Type of loss function, e.g. linear (1), squared (2). Default is 2.}

\item{sigma}{Regularization factor. Default is 0.5.}

\item{...}{Dots are passed to rpart}
}
\value{
Returns a vector with the predictions made by AdaBoost.RTPlus.
}
\description{
AdaBoost.RTPlus
}
\examples{
data(Boston,package="MASS")

idx <- sample(1:nrow(Boston),nrow(Boston)*0.75)
form <- medv ~ .

train <- Boston[idx,]
test <- Boston[-idx,]

preds <- AdaBoost.RTPlus(form,train,test)
}
\references{

}
