% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rtplus.r
\name{AdaBoost.RTPlus.train}
\alias{AdaBoost.RTPlus.train}
\title{SMOTEd AdaBoost.RT+}
\usage{
AdaBoost.RTPlus.train(form, data, niter = 100, thr = 0.01, power = 2,
  sigma = 0.5)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{thr}{Absolute Relative Error threshold used to define prediction cases as hard to predict}

\item{power}{Exponent used to calculate the weight of each decision tree. Common values are 1 (linear), 2 (quadratic, default), 3 (cubic).}

\item{sigma}{Value used for regularization (default 0.5). If no regularization is expected, use NULL}
}
\description{
SMOTEd AdaBoost.RT+
}
\keyword{internal}
