% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada_smote.r
\name{adaSMOTE}
\alias{adaSMOTE}
\title{adaSMOTE}
\usage{
adaSMOTE(form, dat, perc.o, rel.thr = 0.9, k, pc = NULL)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{dat}{A data frame containing the training data}

\item{perc.o}{Over-sampling percentage to apply to randomly selected cases with extreme values}

\item{rel.thr}{Relevance threshold for considering case as having extreme value}

\item{k}{Number of neighbours to use in SMOTE interpolation}

\item{pc}{Relevance function}
}
\description{
adaSMOTE
}
\keyword{internal}
