% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemboost.r
\name{BEMBoost.predict}
\alias{BEMBoost.predict}
\title{predict method}
\usage{
BEMBoost.predict(models, newdata)
}
\arguments{
\item{newdata}{A data frame containing the test data}

\item{model}{A set of decision trees corresponding to the number of boosting iterations}
}
\description{
predict method
}
\keyword{internal}
