% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BEMBoost}
\alias{BEMBoost}
\title{BEMBoost}
\usage{
BEMBoost(form, train, test, t_final = 100, BEM = 0.5, ...)
}
\arguments{
\item{form}{The model formula.}

\item{train}{A data.frame with the training data.}

\item{test}{A data.frame with the test data.}

\item{t_final}{The number of maximum boosting iterations. Default is 100.}

\item{BEM}{Biggest error margin admissible. Defaults to 0.5.}

\item{...}{Dots are passed to rpart}
}
\value{
Returns a vector with the predictions made by BEMBoost.
}
\description{
BEMBoost
}
\examples{
data(Boston,package="MASS")

idx <- sample(1:nrow(Boston),nrow(Boston)*0.75)
form <- medv ~ .

train <- Boston[idx,]
test <- Boston[-idx,]

preds <- BEMBoost(form,train,test)
}
\references{

}
