% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-method.r
\docType{methods}
\name{predict,AdaBoost-method}
\alias{predict,AdaBoost-method}
\title{Predict method for objects of AdaBoost class}
\usage{
\S4method{predict}{AdaBoost}(object, newdata)
}
\arguments{
\item{object}{An AdaBoost object}

\item{newdata}{A data frame containing the test data}
}
\description{
Predict method for objects of AdaBoost class
}
