% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-method.r
\docType{methods}
\name{predict,SmoteBoost-method}
\alias{predict,SmoteBoost-method}
\title{Predict method for objects of SmoteBoost class}
\usage{
\S4method{predict}{SmoteBoost}(object, newdata)
}
\arguments{
\item{object}{A SmoteBoost object}

\item{newdata}{A data frame containing the test data}
}
\description{
Predict method for objects of SmoteBoost class
}
