% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemboost.r
\name{reboost_predict}
\alias{reboost_predict}
\title{predict method}
\usage{
reboost_predict(model, newdata)
}
\arguments{
\item{model}{list of models}

\item{newdata}{new data}
}
\description{
predict method
}
