% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoteboost-class.r
\docType{class}
\name{SmoteBoost-class}
\alias{SmoteBoost-class}
\title{SmoteBoost-class}
\description{
\strong{SmoteBoost} is a S4 class that contains
}
\section{Slots}{

\describe{
\item{\code{model}}{A set of decision trees corresponding to the number of boosting iterations}

\item{\code{form}}{A formula describing the prediction problem}

\item{\code{model_pars}}{Parameters used to create decision trees and guide the boosting process}

\item{\code{modeltype}}{Type of boosting method}

\item{\code{perc_over}}{Over-sampling percentage to apply to randomly selected cases with extreme values.

This package enables the implementation of 5 different proposals for SMOTEBoost described in the work of Moniz et al. (2018), including those based on AdaBoost.R2 (Drucker1997), AdaBoost.RQ, AdaBoost.RT (Solomatine2004), AdaBoost.RT+ (Kankanala2014) and BEMBoost (Feely2000).}
}}

\references{
N. Moniz, R. P. Ribeiro, V. Cerqueira, N. Chawla (2018). "SMOTEBoost for Regression: Improving the Prediction of Extreme Values", in Proceedings of the 5th IEEE International Conference on Data Science and Advanced Analytics, IEEE Xplore. 2018.
H. Drucker, “Improving regressors using boosting techniques”, in Proceedings of the Fourteenth International Conference on Machine Learning, ICML’97, Morgan Kaufmann Publishers Inc., pp. 107–115. 1997.
R. Feely, “Predicting stock market volatility using neural networks", Ph.D. Dissertation, Trinity College, Dublin, 2000.
D. P. Solomatine and D. L. Shrestha, “Adaboost.rt: a boosting algorithm for regression problems,” in 2004 IEEE International Joint Conference on Neural Networks, vol. 2, pp. 1163–1168 vol.2, 2004.
P. Kankanala, S. Das, and A. Pahwa, “Adaboost+: An ensemble learning approach for estimating weather-related outages in distribution systems", IEEE Transactions on Power Systems, vol. 29, no. 1, pp. 359–367, 2014.
}
\keyword{internal}
