% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMOTEBoost.BEM}
\alias{SMOTEBoost.BEM}
\title{SMOTEBoost variant of BEMBoost}
\usage{
SMOTEBoost.BEM(form, train, test, t_final = 100, BEM = 0.5,
  perc.O = 1.5, rel.thr = 0.9, k = 3, coef = 1.5, ...)
}
\arguments{
\item{form}{The model formula.}

\item{train}{A data.frame with the training data.}

\item{test}{A data.frame with the test data.}

\item{t_final}{The number of maximum boosting iterations. Default is 100.}

\item{BEM}{Biggest error margin admissible. Defaults to 0.5.}

\item{perc.O}{Percentage for Oversampling via SMOTE, i.e. percentage of extreme cases to be generated. Default is 1.5.}

\item{rel.thr}{Relevance threshold. Default is 0.9.}

\item{k}{Number of neighbours used in SMOTE. Defaults to 3.}

\item{coef}{Coefficient used in boxplot statistics, which is used to create the relevance function. Default is 1.5.}

\item{...}{Dots are passed to rpart}
}
\value{
Returns a vector with the predictions made by SMOTEBoost.BEM.
}
\description{
SMOTEBoost variant of BEMBoost
}
\examples{
data(Boston,package="MASS")

idx <- sample(1:nrow(Boston),nrow(Boston)*0.75)
form <- medv ~ .

train <- Boston[idx,]
test <- Boston[-idx,]

preds <- SMOTEBoost.BEM(form,train,test)

Snippet
For situations where the ad-hoc big error margin is too large,
This snippet reduces it by half until the BEM error count is more than zero.
}
