% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rq.r
\name{SMOTEBoost.RQ}
\alias{SMOTEBoost.RQ}
\title{SMOTEd AdaBoost.RQ}
\usage{
SMOTEBoost.RQ(form, data, niter = 100, power = 2, perc.O = 1.5)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{niter}{boosting iters}

\item{power}{pw}

\item{perc.O}{po}
}
\description{
SMOTEd AdaBoost.RQ
}
