% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rq.r
\name{SMOTEBoost.RQ.train}
\alias{SMOTEBoost.RQ.train}
\title{SMOTEd AdaBoost.RQ}
\usage{
SMOTEBoost.RQ.train(form, data, niter = 100, power = 2, perc.O = 1.5)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{power}{Exponent used to calculate the weight of each decision tree. Common values are 1 (linear), 2 (quadratic, default), 3 (cubic).}

\item{perc.O}{po}
}
\description{
SMOTEd AdaBoost.RQ
}
\keyword{internal}
