% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rt.r
\name{SMOTEBoost.RT}
\alias{SMOTEBoost.RT}
\title{SMOTEd AdaBoost.RT}
\usage{
SMOTEBoost.RT(form, data, niter = 100, thr = 0.1, power = 2,
  perc.O = 1.5)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{niter}{boosting iters}

\item{thr}{thr}

\item{power}{pw}

\item{perc.O}{pwe}
}
\description{
SMOTEd AdaBoost.RT
}
