% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rt.r
\name{SMOTEBoost.RT.train}
\alias{SMOTEBoost.RT.train}
\title{Create a SMOTEd AdaBoost.RT model}
\usage{
SMOTEBoost.RT.train(form, data, niter = 100, thr = 0.1, power = 2,
  perc.O = 1.5)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{thr}{Absolute Relative Error threshold used to define prediction cases as hard to predict}

\item{power}{Exponent used to calculate the weight of each decision tree. Common values are 1 (linear), 2 (quadratic, default), 3 (cubic).}

\item{perc.O}{Over-sampling percentage to apply to randomly selected cases with extreme values}
}
\description{
Create a SMOTEd AdaBoost.RT model
}
\keyword{internal}
