% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rtplus.r
\name{SMOTEBoost.RTPlus.train}
\alias{SMOTEBoost.RTPlus.train}
\title{SMOTEd AdaBoost.RT+}
\usage{
SMOTEBoost.RTPlus.train(form, data, niter = 100, thr = 0.01, power = 2,
  sigma = 0.5, perc.O = 1.5)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{niter}{boosting iterations}

\item{thr}{thr}

\item{power}{pw}

\item{sigma}{sad}

\item{perc.O}{perco}
}
\description{
SMOTEd AdaBoost.RT+
}
