% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemboost.r
\name{SMOTEBoostBEM.train}
\alias{SMOTEBoostBEM.train}
\title{Create a SMOTEd BEMBoost model}
\usage{
SMOTEBoostBEM.train(form, data, niter = 100, BEM = 0.5, perc.O = 1.5)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{BEM}{Big Error Margin. A value that defines which errors (absolute distance) should be considered as being big}

\item{perc.O}{Over-sampling percentage to apply to randomly selected cases with extreme values}
}
\description{
Create a SMOTEd BEMBoost model
}
\keyword{internal}
