% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost-class.r
\docType{class}
\name{AdaBoost-class}
\alias{AdaBoost-class}
\title{AdaBoost-class}
\description{
\strong{AdaBoost} is a S4 class that contains
}
\section{Slots}{

\describe{
\item{\code{model}}{A set of decision trees corresponding to the number of boosting iterations}

\item{\code{form}}{A formula describing the prediction problem}

\item{\code{model_pars}}{Parameters used to create decision trees and guide the boosting process}

\item{\code{modeltype}}{Type of boosting method}
}}

\keyword{internal}
