% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_r2.r
\name{AdaBoost.R2.predict}
\alias{AdaBoost.R2.predict}
\title{AdaBoost.R2 predict method}
\usage{
AdaBoost.R2.predict(models, newdata)
}
\arguments{
\item{models}{list of models}

\item{newdata}{newdata}
}
\description{
AdaBoost.R2 predict method
}
