% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_r2.r
\name{AdaBoost.R2.train}
\alias{AdaBoost.R2.train}
\title{AdaBoost.R2}
\usage{
AdaBoost.R2.train(form, data, niter = 100, power = 2)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{niter}{boosting iters}

\item{power}{power}
}
\description{
AdaBoost.R2
}
