% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rq.r
\name{AdaBoost.RQ.predict}
\alias{AdaBoost.RQ.predict}
\title{AdaBoost.RQ predict method}
\usage{
AdaBoost.RQ.predict(models, betas, newdata)
}
\arguments{
\item{models}{list of models}

\item{betas}{betas}

\item{newdata}{newdata}
}
\description{
AdaBoost.RQ predict method
}
