% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rt.r
\name{AdaBoost.RT.predict}
\alias{AdaBoost.RT.predict}
\title{AdaBoost.RT predict method}
\usage{
AdaBoost.RT.predict(models, newdata, betas)
}
\arguments{
\item{models}{list of models}

\item{newdata}{newdata}

\item{betas}{betas}
}
\description{
AdaBoost.RT predict method
}
