% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rt.r
\name{AdaBoost.RT.train}
\alias{AdaBoost.RT.train}
\title{AdaBoost.RT}
\usage{
AdaBoost.RT.train(form, data, niter = 100, thr = 0.1, power = 2)
}
\arguments{
\item{form}{formaula}

\item{data}{trianing data}

\item{niter}{boosting ites}

\item{thr}{dd}

\item{power}{pw}
}
\description{
AdaBoost.RT
}
