% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_rtplus.r
\name{AdaBoost.RTPlus.predict}
\alias{AdaBoost.RTPlus.predict}
\title{predict method for AdaBoost.RTPlus}
\usage{
AdaBoost.RTPlus.predict(models, newdata, delta)
}
\arguments{
\item{models}{lsit of models}

\item{newdata}{new data}

\item{delta}{delta}
}
\description{
predict method for AdaBoost.RTPlus
}
