% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ada_smote.r
\name{adaSMOTE}
\alias{adaSMOTE}
\title{adaSMOTE}
\usage{
adaSMOTE(form, dat, perc.o, rel.thr, k, pc = NULL)
}
\arguments{
\item{form}{formula}

\item{dat}{training data}

\item{perc.o}{percentage of oversmapling????}

\item{rel.thr}{???}

\item{k}{??}

\item{pc}{???}
}
\description{
adaSMOTE
}
