% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{adaSMOTE}
\alias{adaSMOTE}
\title{Title}
\usage{
adaSMOTE(form, dat, perc.o, rel.thr, k, pc = NULL)
}
\arguments{
\item{form}{The model formula.}

\item{dat}{A data.frame with the training data.}

\item{perc.o}{Percentage for Oversampling via SMOTE, i.e. percentage of extreme cases to be generated. Default is 1.5.}

\item{rel.thr}{Relevance threshold. Default is 0.9.}

\item{k}{Number of neighbours used in SMOTE. Defaults to 3.}

\item{pc}{Relevance function phi.}
}
\description{
Title
}
