% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemboost.r
\name{BEMBoost.train}
\alias{BEMBoost.train}
\title{BEMBoost}
\usage{
BEMBoost.train(form, data, niter = 100, BEM = 0.5)
}
\arguments{
\item{form}{formual}

\item{data}{trianing data}

\item{niter}{boosting iters}

\item{BEM}{??}
}
\description{
BEMBoost
}
