% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost-class.r
\docType{methods}
\name{predict,AdaBoost-method}
\alias{predict,AdaBoost-method}
\title{Predict method for objects of AdaBoost class}
\usage{
\S4method{predict}{AdaBoost}(object, newdata)
}
\arguments{
\item{object}{An AdaBoost object}

\item{newdata}{A data frame containing the test data}
}
\description{
Method for prediction when using objects of the class AdaBoost.
}
\examples{
library(ReBoost)
data(iris)
ind <- sample(0.75*nrow(iris))
tr <- iris[ind,]
ts <- iris[-ind,]
form<-Sepal.Length ~ .

#For AdaBoost.RQ
m1 <- AdaBoost(form, tr, modeltype = "RQ", model_pars = NULL); predict(m1, ts)
m1p <- AdaBoost(form, tr, modeltype = "RQ", model_pars = list(niter=100,power=2)); predict(m1p, ts)

}
