% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoteboost-class.r
\docType{methods}
\name{predict,SMOTEBoost-method}
\alias{predict,SMOTEBoost-method}
\title{Predict method for objects of SMOTEBoost class}
\usage{
\S4method{predict}{SMOTEBoost}(object, newdata)
}
\arguments{
\item{object}{A SMOTEBoost object}

\item{newdata}{A data frame containing the test data}
}
\description{
Predict method for objects of SMOTEBoost class
}
\examples{
library(ReBoost)
data(iris)
ind <- sample(0.75*nrow(iris))
tr <- iris[ind,]
ts <- iris[-ind,]
form<-Sepal.Length ~ .

#For SMOTEBoost.RQ
m1 <- SMOTEBoost(form, tr, modeltype = "RQ", model_pars = NULL); predict(m1, ts)
m1p <- SMOTEBoost(form, tr, modeltype = "RQ", model_pars = list(niter=100,power=2), perc_over=1); predict(m1p, ts)

}
