% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_r2.r
\name{SMOTEBoost.R2.train}
\alias{SMOTEBoost.R2.train}
\title{SMOTEd AdaBoost.R2}
\usage{
SMOTEBoost.R2.train(form, data, niter = 100, power = 2, perc.O = 1.5)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{niter}{no iterations}

\item{power}{power?}

\item{perc.O}{per}
}
\description{
SMOTEd AdaBoost.R2
}
