% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_r2.r
\name{SMOTEBoost.R2.train}
\alias{SMOTEBoost.R2.train}
\title{SMOTEd AdaBoost.R2}
\usage{
SMOTEBoost.R2.train(form, data, niter = 100, power = 2, perc.O = 1.5)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the training data}

\item{niter}{The number of boosting iterations}

\item{power}{Exponent used to calculate the weight of each decision tree. Common values are 1 (linear), 2 (quadratic, default), 3 (cubic).}

\item{perc.O}{Over-sampling percentage to apply to randomly selected cases with extreme values}
}
\description{
SMOTEd AdaBoost.R2
}
\keyword{internal}
