% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bemboost.r
\name{SMOTEBoostBEM.train}
\alias{SMOTEBoostBEM.train}
\title{SMOTEd BEMBoost}
\usage{
SMOTEBoostBEM.train(form, data, niter = 100, BEM = 0.5, perc.O = 1.5)
}
\arguments{
\item{form}{form}

\item{data}{train data}

\item{niter}{boost iters}

\item{BEM}{??}

\item{perc.O}{??}
}
\description{
SMOTEd BEMBoost
}
